move seed data to consts
